ogv-frontend/src/app/modal/modal.component.ts
import { Component, OnInit } from '@angular/core';
import { MdbModalRef } from 'mdb-angular-ui-kit/modal';
import { ImageService } from 'src/services/image.service';

@Component({
  selector: 'app-modal',
  templateUrl: './modal.component.html',
  styleUrls: ['./modal.component.scss']
})

export class ModalComponent {
  title: string | null = null;
  id: number | null = null;

  constructor(public modalRef: MdbModalRef<ModalComponent>) {}

  close(): void {
    const closeMessage = 'Modal closed';
    this.modalRef.close(closeMessage)
    this.refresh();
  }

  refresh(): void {
    window.location.reload();
  }
  // providers: [ImageService]

}
